refactor: Integer로 설정되어 있던 Key value Long으로 수정
Integer로 설정되어 있던 Key value Long으로 수정

+++ backend/Exhale/src/main/java/com/ssafy/exhale/service/ArticleService.java
@@ -19,7 +19,7 @@ public class ArticleService {
         return articleRepository.findAllByBoardId(boardId, pageRequest);
     }
 
-    public Article getArticle(int articleId){
+    public Article getArticle(Long articleId){
         Article article = null;
         if(articleRepository.findById(articleId).isPresent()){
             article = articleRepository.findById(articleId).get();
